# back-end/schema/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Personal trading account API
servers:
  - url: http://localhost:3001/api
paths:
...
  /user/editProfile:
    post:
      tags:
        - user
      summary: Edit user profile
      description: Edit user profile
      operationId: editProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/definitions/User'
            example:
              login: Andrew
              password: '666666'
        description: User credentials
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User profile edited'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
  /user/deleteUser:
    delete:
      tags:
        - user
      summary: Delete user
      description: Delete user
      operationId: deleteUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User deleted'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
...
  /user/getUsers:
    get:
      tags:
        - user
      summary: Get all users
      description: Get all users
      operationId: getUsers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Users found'
                  data:
                    $ref: '#/definitions/User'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
  /user/getUsers/{userId}:
    get:
      tags:
        - user
      summary: Get user by id
      description: Get user by id
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: User id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User found'
                  data:
                    $ref: '#/definitions/User'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User not found'
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
security:
  - cookieAuth: []
definitions:
  User:
    required:
      - login
      - password
    type: object
    properties:
      _id:
        type: ObjectId
        format: _id
        example: 63080201e3957808e2bebe61
      type:
        type: string
        default: User
      groupId:
        type: ObjectId
        format: _id
        example: 5445654das654sada465s4ad
        ref: Group
        default: null
      login:
        type: string
        example: user1
        unique: true
      password:
        type: string
        example: 123456
      permission:
        type: string
        default: user
      isAdmin:
        type: boolean
        default: false
      isBanned:
        type: boolean
        default: false
      subscriptions:
        type: array
        items:
          type: string
          example: Group
      username:
        type: string
        example: User1
      steamId:
        type: string
        example: 76561198000000000
      avatar:
        type: string
        example: https://steamcdn-a.akamaihd.net/steamcommunity/public/images/avatars/00/00000000000000000.jpg
      created:
        type: number
        default: 1663117006355
      updated:
        type: number
        default: 1663117006355
    xml:
      name: user
...